Split up `git::worktree`
